requirements.txt
PyYAML
selenium
pypdf
openai
requests
pyautogui
webdriver_manager
pyrebase4
sseclient-py